dev-requirements.txt
autobahn==23.6.2
Automat==22.10.0
black==24.3.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coverage==7.4.4
crispy-bootstrap5==2024.2
example-package==0.0.1
factory-boy==3.3.0
Faker==24.4.0
h11==0.14.0
hyperlink==21.0.0
incremental==22.10.0
iniconfig==2.0.0
joblib==1.3.2
mypy-extensions==1.0.0
outcome==1.3.0.post0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyPDF2==3.0.1
pypng==0.20220715.0
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-base-url==2.1.0
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.14.0
pytest-selenium==4.1.0
pytest-variables==3.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
tenacity==8.2.3
threadpoolctl==3.4.0
trio==0.25.0
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
wcwidth==0.2.13
wheel==0.43.0
wsproto==1.2.0
